requirements_airflow.txt
webdriver-manager
selenium
tqdm
pydantic
pandas
boto3
apache-airflow
apache-airflow-providers-apache-spark
# psycopg2-binary
# asyncpg